Some suggested changes for nested dataframe creation and repo layout
